fix: mmr type to int

+++ include/model/Player.h
@@ -18,16 +18,16 @@ private:
     std::string playerId;
     int seedingRank;
     Race race;
-    double mmr;
+    int mmr;
     std::vector<int> prefList;
 
 public:
-    Player(std::string id, int rank, Race r, double m);
+    Player(std::string id, int rank, Race r, int m);
 
     std::string getPlayerId() const;
     int getSeedingRank() const;
     Race getRace() const;
-    double getMMR() const;
+    int getMMR() const;
     const std::vector<int>& getPrefList() const;
 
     void setPrefList(std::vector<int> list);
